servers/AI-Services-API/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
py-eureka-client==0.13.2
pydantic==2.10.5
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.20
pandas>=2.0.0
numpy>=1.24.0,&lt;2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
openai>=1.0.0
faiss-cpu>=1.7.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
